Jarvis/requirements.txt
PyQt5
python-dotenv
requests
googlesearch-python
groq
google-generativeai
selenium
webdriver-manager
mtranslate
pygame
edge-tts
cohere
rich
AppOpener
pywhatkit
beautifulsoup4
keyboard
pyautogui
pygetwindow
